Extract drag button state updater to separate file. Enable zooming while drag button is active.
